WIP: Aligning LendBike and ReturnBike
