--- src/modules/bicycles/pages/components/SelectBikePage/SelectBikePage.styles.ts
@@ -1,44 +1,0 @@
-import { makeStyles } from '@material-ui/core';
-
-export default makeStyles(theme => ({
-  cardPosition: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  root: {
-    display: 'flex',
-    marginBottom: '20px',
-    width: 600,
-    cursor: 'pointer',
-  },
-  details: {
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  content: {
-    flex: '1 0 auto',
-  },
-  cover: {
-    width: 151,
-  },
-  controls: {
-    display: 'flex',
-    alignItems: 'center',
-    paddingLeft: theme.spacing(1),
-    paddingBottom: theme.spacing(1),
-  },
-  playIcon: {
-    height: 38,
-    width: 38,
-  },
-  pageNameStyle: {
-    textAlign: 'center',
-    textTransform: 'uppercase',
-    fontWeight: 500,
-    paddingBottom: '18px',
-  },
-  titleStyle: {
-    paddingBottom: '20px',
-  },
-}));
